fix(scroll): fix scroll button
